cloud_functions/get_rates/requirements.txt
datetime
requests
google-cloud-storage
pytest
